Cleaned up negative pattern generation of stringly typed params
