+++ core/src/main/kotlin/in/specmatic/core/pattern/TabularPattern.kt
@@ -139,112 +139,6 @@ fun newBasedOn(patternMap: Map<String, Pattern>, row: Row, resolver: Resolver): 
     return patternList(patternCollection)
 }
 
-fun negativeBasedOn(patternMap: Map<String, Pattern>, row: Row, resolver: Resolver, stringlyCheck: Boolean=false): List<Map<String, Pattern>> {
-    val eachKeyMappedToPatternMap = patternMap.mapValues { patternMap }
-    val negativePatternsMap = getNegativePatterns(patternMap, resolver, stringlyCheck, row)
-
-    val modifiedPatternMap: Map<String, List<Map<String, List<Pattern>>>> = eachKeyMappedToPatternMap.mapValues { (keyToNegate, patterns) ->
-        val negativePatterns = negativePatternsMap[keyToNegate]
-        negativePatterns!!.map { negativePattern ->
-            patterns.mapValues { (key, pattern) ->
-                attempt(breadCrumb = key) {
-                    when (key == keyToNegate) {
-                        true ->
-                            attempt(breadCrumb = "Setting $key to $negativePattern for negative test scenario") {
-                                if (stringlyCheck && patternIsEnum(negativePattern, resolver)) {
-                                    negativeBasedOnForEnum(negativePattern)
-                                } else
-                                    newBasedOn(Row(), key, negativePattern, resolver)
-                            }
-                        else -> newBasedOn(row, key, pattern, resolver)
-                    }
-                }
-            }
-        }
-    }
-    if (modifiedPatternMap.values.isEmpty())
-      return listOf(emptyMap())
-    return modifiedPatternMap.values.map { list: List<Map<String, List<Pattern>>> ->
-        list.toList().map { patternList(it) }.flatten()
-    }.flatten()
-}
-
-private fun getNegativePatterns(
-    patternMap: Map<String, Pattern>,
-    resolver: Resolver,
-    stringlyCheck: Boolean,
-    row: Row
-): Map<String, List<Pattern>> {
-    return patternMap.mapValues { (key, pattern) ->
-        val resolvedPattern = resolvedHop(pattern, resolver)
-        if(stringlyCheck) {
-            if (patternIsEnum(resolvedPattern, resolver)) {
-                shortCircuitStringlyEnumGenerationToOneEnumValue(resolvedPattern, resolver)
-            } else {
-                resolvedPattern
-                    .negativeBasedOn(row.stepDownOneLevelInJSONHierarchy(withoutOptionality(key)), resolver)
-                    .filterNot {
-                        isStringly(resolvedPattern, it, resolver)
-                    }.filterNot {
-                        it is NullPattern
-                    }
-            }
-        } else {
-            resolvedPattern.negativeBasedOn(row.stepDownOneLevelInJSONHierarchy(withoutOptionality(key)), resolver)
-        }
-        // TODO: Refactor special handling of scalar and array query params when the stringly check is refactored
-//        if (stringlyCheck && (resolvedPattern is StringPattern || isStringBasedQueryParameterScalarPattern(
-//                resolvedPattern
-//            ) || isStringBasedQueryParameterArrayPattern(resolvedPattern))
-//        ) {
-//            emptyList()
-//        } else if (stringlyCheck && isScalar(resolvedPattern)) {
-//            resolvedPattern.negativeBasedOn(row.stepDownOneLevelInJSONHierarchy(withoutOptionality(key)), resolver)
-//                .filterNot { it is NullPattern }
-//        } else if (stringlyCheck && patternIsEnum(resolvedPattern, resolver)) {
-//            shortCircuitStringlyEnumGenerationToOneEnumValue(resolvedPattern, resolver)
-//        } else {
-//            resolvedPattern.negativeBasedOn(row.stepDownOneLevelInJSONHierarchy(withoutOptionality(key)), resolver)
-//        }
-    }
-}
-
-private fun isStringly(
-    resolvedPattern: Pattern,
-    it: Pattern,
-    resolver: Resolver
-) = resolvedPattern.matches(it.generate(resolver).toStringValue(), resolver) is Result.Success
-
-private fun isStringBasedQueryParameterArrayPattern(resolvedPattern: Pattern) =
-    resolvedPattern is QueryParameterArrayPattern &&  resolvedHop(resolvedPattern.pattern.first(), Resolver()) is StringPattern
-
-private fun isStringBasedQueryParameterScalarPattern(resolvedPattern: Pattern) =
-    resolvedPattern is QueryParameterScalarPattern && resolvedHop(resolvedPattern.pattern, Resolver()) is StringPattern
-
-private fun negativeBasedOnForEnum(pattern: Pattern): List<Pattern> {
-    val enumPattern = (pattern as EnumPattern).pattern
-    val firstEnumOption = enumPattern.pattern.first() as ExactValuePattern
-    val valueOfFirstEnumOption = firstEnumOption.pattern
-    val patternOfFirstValue = valueOfFirstEnumOption.type()
-    return listOf(patternOfFirstValue)
-}
-
-private fun shortCircuitStringlyEnumGenerationToOneEnumValue(
-    pattern: Pattern,
-    resolver: Resolver
-): List<AnyPattern> {
-    val resolvedAnyPattern = (resolvedHop(pattern, resolver) as EnumPattern).pattern
-    val firstEnumValue = resolvedAnyPattern.pattern.first() as ExactValuePattern
-
-    return listOf(AnyPattern(listOf(firstEnumValue)))
-}
-
-fun patternIsEnum(pattern: Pattern, resolver: Resolver): Boolean {
-    val resolvedPattern = resolvedHop(pattern, resolver)
-
-    return resolvedPattern is EnumPattern
-}
-
 fun newBasedOn(patternMap: Map<String, Pattern>, resolver: Resolver): List<Map<String, Pattern>> {
     val patternCollection = patternMap.mapValues { (key, pattern) ->
         attempt(breadCrumb = key) {
@@ -254,29 +148,6 @@ fun newBasedOn(patternMap: Map<String, Pattern>, resolver: Resolver): List<Map<S
 
     return patternValues(patternCollection)
 }
-
-fun negativeBasedOn(patternMap: Map<String, Pattern>, resolver: Resolver, stringlyCheck:Boolean=false): List<Map<String, Pattern>> {
-    val patternCollection = patternMap.mapValues { (key, pattern) ->
-        attempt(breadCrumb = key) {
-            val resolvedPattern = resolvedHop(pattern, resolver)
-            if(stringlyCheck && resolvedPattern is StringPattern) {
-                emptyList()
-            }
-            else if (stringlyCheck && isScalar(resolvedPattern)) {
-                negativeBasedOn(key, resolvedPattern, resolver).filterNot { it is NullPattern  }
-            }
-            else {
-                negativeBasedOn(key, resolvedPattern, resolver)
-            }
-        }
-    }
-
-    return patternValues(patternCollection)
-}
-
-private fun isScalar(resolvedPattern: Pattern) =
-    resolvedPattern is ScalarType
-        || (resolvedPattern is QueryParameterScalarPattern && resolvedPattern.pattern is ScalarType)
 
 fun newBasedOn(row: Row, key: String, pattern: Pattern, resolver: Resolver): List<Pattern> {
     val keyWithoutOptionality = key(pattern, key)
@@ -327,14 +198,6 @@ fun newBasedOn(row: Row, key: String, pattern: Pattern, resolver: Resolver): Lis
 }
 
 fun newBasedOn(key: String, pattern: Pattern, resolver: Resolver): List<Pattern> {
-    return resolver.withCyclePrevention(pattern, isOptional(key)) { cyclePreventedResolver ->
-        pattern.newBasedOn(cyclePreventedResolver)
-    }?:
-    // Handle cycle (represented by null value) by using empty list for optional properties
-    listOf()
-}
-
-fun negativeBasedOn(key: String, pattern: Pattern, resolver: Resolver): List<Pattern> {
     return resolver.withCyclePrevention(pattern, isOptional(key)) { cyclePreventedResolver ->
         pattern.newBasedOn(cyclePreventedResolver)
     }?:
